requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
kiteconnect==4.1.0
httpx==0.25.1
python-multipart==0.0.6
jinja2==3.1.2 